+++ spl-net/src/main/java/bgu/spl/net/api/Bidi/Messages/ErrorMessage.java
@@ -4,7 +4,7 @@ import bgu.spl.net.api.Bidi.BidiMessagingProtocol;
 import bgu.spl.net.api.Bidi.BidiMessagingProtocolImp;
 
 public class ErrorMessage implements Message{
-    private short opcode = 11;
+    private short opcode = (short)11;
     private short msgOpcode;
 
     public ErrorMessage(short _msgOpcode){
